images_downloader/requirements.txt
prettytable == 3.6.0
requests == 2.28.2
aiohttp
aiofiles
scipy == 1.10.1
pillow == 9.4.0
numpy == 1.24.2
